added run test command

+++ src/Database/Migration/ColumnDefination.php
@@ -136,6 +136,11 @@ class ColumnDefination
 
     public function dropIndex($name)
     {
+        if ($this->driver == "sqlite") {
+            (new Schema)->query(concat(["ALTER TABLE `$this->table` DROP IF EXISTS INDEX", "`$name`"]));
+            return null;
+        }
+
         (new Schema)->query(concat(["ALTER TABLE `$this->table` DROP INDEX", "`$name`"]));
     }
